requirements.txt
pymongo==2.6.2
numpy==1.9.0
pandas==0.23.0
scipy==0.18.1
python-slugify==0.0.7
Unidecode==0.04.12